requirements.txt
streamlit>=1.29
opencv-python-headless>=4.8.0.76
opencv-contrib-python-headless>=4.8.0.76
numpy>=1.24
Pillow>=9.5.0
